Add shortcut to revert last commit without staging
Sometimes, I want to undo some (but not all) of the changes made in a
previous commit. This shortcut allows me to undo the changes specified
by that last commit, then stage or stash changes as desired.

`git revert --no-commit HEAD` will have the side effect of staging the
reverted changes. Also, notably -- "the revert is done against the
beginning state of your index."