Change git sed to avoid unnecessary writes
Previously, I would invoke the custom command like so:

    git sed 's/improt/import/g'

The command would identify all files tracked in Git and pipe my
expression directly to `sed -i`.

However, the key problem with this is that `sed -i` will overwrite
symlinks, update timestamps (even if there's no substitution to make),
and add newlines to DOS files. This can be problematic in large
repositories, especially when binary content is tracked.

Since 99% of my sed usage is just simple substitution of a global
string, I've rewritten the alias to be just `git sed improt import`. If
I want to return to the full power of `sed -i`, I could employ a
strategy like so:

1. Parse the search pattern from the sed expression
2. Identify files that have the search string (as we do now)
3. Pipe only files that need substitution through `sed -i`