Ignore submodules when deciding to rebase with fix
When I invoke `git fix <hash>`, I want to trigger an immediate rebase of
a fixup commit, but only if it's safe to do so. If there are any staged
or unstaged changes to tracked files, `git rebase` will warn that we
cannot proceed. This is why I check the exit code of `git diff` before
trying to rebase (I'd rather just make the commit & trigger the rebase
later).

However, changes to submodules are fine to leave as-is! `git rebase`
won't complain, since those changes are tracked separately.