Fetch from all remote branches
When working with repositories that have two remotes (e.g. a personal
clone of an open-source project), it's often helpful to fetch both
`origin` (or whatever the default remote is) *and* any other remotes.

This takes a bit longer when we have remotes we don't care to fetch, but
most of my repos have a single remote anyway, and nothing stops me from
doing a plain `git fetch <remote>`.